requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyjwt==2.8.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.3
sqlalchemy-utils==0.41.0

# AI/ML
langchain==0.1.0
openai==1.3.5
sentence-transformers==2.2.2
huggingface-hub==0.16.4
numpy==1.26.1
torch>=2.0.0
transformers>=4.30.0
langchain-community==0.0.20
langchain-core==0.1.23

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.2
python-magic==0.4.27

# Utils
pydantic==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2
